NPMLE for Logistic-CoxPH Cure-Rate Model

NPMLE for Logistic-CoxPH Cure-Rate Model

PO(formula, data, C, df, weights, subset, init,control,
singular.ok = TRUE,model = FALSE,x = FALSE, y = TRUE, tt,
method = c('U-method','B-spline','NPMLE','glasso','glasso-PLH'),...)

Arguments

formula

a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv.cure function.

data

a data.frame in which to interpret the variables named in the formula, or in the subset and the weights argument.

C

Censoring time

df

degree of freedom

weights

vector of case weights, see the note below. For a thorough discussion of these see the book by Therneau and Grambsch.

subset

expression indicating which subset of the rows of data should be used in the fit. All observations are included by default.

init

vector of initial values of the iteration. Default initial value is zero for all variables.

control

Object of class PO.control specifying iteration limit and other control options. Default is PO.control(...).

singular.ok

logical value indicating how to handle collinearity in the model matrix. If TRUE, the program will automatically skip over columns of the X matrix that are linear combinations of earlier columns. In this case the coefficients for such columns will be NA, and the variance matrix will contain zeros. For ancillary calculations, such as the linear predictor, the missing coefficients are treated as zeros.

model

logical value: if TRUE, the model frame is returned in component model.

x

logical value: if TRUE, the x matrix is returned in component x.

y

logical value: if TRUE, the response vector is returned in component y.

tt

optional list of time-transform functions.

method

a character string specifying the method in c('U-method', 'B-spline','NPMLE') for estimation. The default method is the U-method.

...

other parameters passed to PO.control.

Author

Jue Hou

Examples

 # A simulated data set
require(survival)
data('sim_PO_data')
res = PO(Surv(X, delta) ~ Z[,1]+ Z[,2]+ Z[,3],data = sim_PO_data)
#>    [1] 0.8748721 0.9928889 0.9102891 0.8793472 0.9693538 0.8019397 0.9969686
#>    [8] 0.9231711 0.9887244 0.6634584 0.7837635 0.9989950 0.5533004 0.9317363
#>   [15] 0.8609432 0.8131668 0.9604049 0.9178243 0.5027985 0.9563075 0.9855177
#>   [22] 0.6550067 0.9643190 0.6463882 0.6239506 0.9681142 0.7262641 0.9717863
#>   [29] 0.9617388 0.9822393 0.9617388 0.9383299 0.5257901 0.9159626 0.7115348
#>   [36] 0.8793472 0.9159626 0.9617388 0.9681142 0.8019397 0.9717863 0.9159626
#>   [43] 0.8075972 0.9989950 0.9681142 0.5027985 0.9446998 0.8537079 0.9446998
#>   [50] 0.9231711 0.8414829 0.8131668 0.9576995 0.9908225 0.8185791 0.9655996
#>   [57] 0.9729890 0.5027985 0.5027985 0.9350404 0.6463882 0.8075972 0.9446998
#>   [64] 0.9617388 0.9833517 0.7961472 0.9833517 0.7369971 0.8047734 0.7773915
#>   [71] 1.0000000 0.7990959 0.7439677 0.9576995 0.9520570 0.5027985 0.9064278
#>   [78] 0.9549011 0.5027985 0.9005405 0.9576995 0.9969686 0.9887244 0.9876578
#>   [85] 0.9534802 0.9717863 0.9908225 0.9491894 0.8923287 0.5027985 0.9753307
#>   [92] 0.9549011 0.9969686 0.9844383 0.7961472 0.9446998 0.9064278 0.9865912
#>   [99] 0.5027985 0.8923287 0.7837635 0.6843811 0.5027985 1.0000000 0.8464040
#>  [106] 0.9887244 0.8985304 0.9178243 0.9949319 0.9249129 0.8131668 0.9520570
#>  [113] 0.8609432 0.5691607 0.5201364 0.5027985 0.9604049 0.9969686 0.9833517
#>  [120] 0.7990959 0.9102891 0.8679887 0.5797007 0.6923162 0.9446998 0.9266385
#>  [127] 0.9549011 0.7837635 0.9383299 0.5027985 0.9178243 0.9729890 0.9693538
#>  [134] 0.6000033 0.9799603 0.6550067 0.6286070 0.7961472 0.7077701 0.8047734
#>  [141] 0.8837221 0.9617388 0.7708034 0.9928889 0.9655996 0.7405234 0.8985304
#>  [148] 0.9630332 0.8633149 0.9383299 0.9159626 0.6331953 0.9799603 0.7001390
#>  [155] 0.5027985 0.8314871 0.9908225 0.9576995 0.9266385 0.9462113 0.7405234
#>  [162] 0.7039649 0.5368593 0.8748721 0.9729890 0.5479286 0.7773915 0.8339992
#>  [169] 0.7541759 0.9520570 0.9462113 0.9668585 0.9399311 0.9681142 0.9655996
#>  [176] 0.9617388 0.9822393 0.9590579 0.9140740 0.9655996 0.5027985 0.9576995
#>  [183] 0.9693538 0.9520570 0.9643190 0.6761095 0.9897774 0.8264171 0.8633149
#>  [190] 0.7837635 0.8923287 0.9178243 0.9604049 0.9446998 0.9446998 0.5027985
#>  [197] 0.9617388 0.6802830 0.8264171 0.9822393 0.8837221 0.9604049 0.9693538
#>  [204] 1.0000000 0.7001390 0.9979828 0.9811011 0.9741627 0.9822393 0.9317363
#>  [211] 0.8158864 0.7262641 0.5027985 0.5848308 0.8858934 0.5027985 0.9617388
#>  [218] 0.6098049 0.9563075 0.8923287 0.8748721 0.9102891 0.9844383 0.5027985
#>  [225] 0.9383299 0.9446998 0.9928889 0.9025417 0.5027985 0.5027985 0.6286070
#>  [232] 0.8185791 0.9576995 0.9477132 0.7001390 0.9705700 0.9178243 0.9822393
#>  [239] 0.9729890 0.9491894 0.7262641 0.9822393 0.5797007 0.9491894 0.9729890
#>  [246] 0.7868986 0.7961472 0.9044868 0.9865912 0.9446998 0.9617388 0.9576995
#>  [253] 0.9266385 0.9969686 0.9844383 0.9717863 0.9918557 0.5027985 0.5027985
#>  [260] 0.9855177 0.9300484 0.9491894 0.7474120 0.8158864 0.9928889 0.9446998
#>  [267] 0.9044868 0.6050033 0.9617388 0.5027985 0.9908225 0.9317363 0.9897774
#>  [274] 0.7077701 0.9366866 0.7773915 0.5027985 0.8238264 0.6376233 0.9383299
#>  [281] 0.8944039 0.7001390 0.9590579 0.7773915 0.9178243 0.8264171 0.7900212
#>  [288] 0.9025417 0.7575278 0.9563075 0.9630332 0.9576995 0.9617388 0.9491894
#>  [295] 0.9317363 0.9159626 0.5950033 0.8390007 0.8488573 0.9617388 0.9300484
#>  [302] 0.6802830 1.0000000 0.6843811 0.8656802 0.9249129 0.6923162 0.9717863
#>  [309] 0.6098049 0.6331953 0.8656802 0.5027985 0.8726056 0.9643190 0.5027985
#>  [316] 0.9693538 0.9681142 0.5027985 0.9231711 0.7641728 0.9833517 0.9776599
#>  [323] 0.8985304 0.9317363 0.7369971 0.9178243 0.9693538 0.9729890 0.9249129
#>  [330] 0.5027985 0.9159626 0.9266385 0.6463882 0.9833517 0.9491894 0.9833517
#>  [337] 0.9844383 0.8185791 0.6802830 0.8365037 0.7931315 0.6286070 0.5027985
#>  [344] 0.9729890 0.8985304 0.6331953 0.5027985 0.8726056 0.9446998 0.9630332
#>  [351] 0.9317363 0.7262641 0.8264171 0.7474120 0.9617388 0.6420207 0.7773915
#>  [358] 0.9729890 0.8585517 0.9446998 0.8390007 0.7990959 0.5027985 0.9477132
#>  [365] 0.9681142 0.5744307 0.9534802 0.9446998 0.5027985 0.5027985 0.9178243
#>  [372] 0.9655996 0.7990959 0.9989950 0.5027985 0.9083605 0.9196170 0.9446998
#>  [379] 0.9415296 0.9462113 0.5423940 0.9630332 0.6331953 0.7931315 0.8512826
#>  [386] 0.9788155 0.8075972 0.9283481 0.8748721 0.5479286 0.8238264 0.7405234
#>  [393] 0.8815401 0.6331953 0.9178243 0.9317363 0.8185791 0.9415296 0.6592325
#>  [400] 0.9300484 0.9617388 0.9025417 0.7575278 0.6420207 0.7152601 0.9822393
#>  [407] 0.9491894 0.9213957 0.7931315 0.9590579 0.8748721 0.9462113 0.9491894
#>  [414] 0.9446998 0.6239506 0.8923287 0.5027985 0.9717863 0.5027985 0.9604049
#>  [421] 0.9317363 0.5027985 0.8238264 0.8314871 0.9959503 0.9833517 0.8561332
#>  [428] 0.9693538 0.8793472 0.9506232 0.9383299 0.5899609 0.5027985 0.8131668
#>  [435] 0.9989950 0.8771097 0.7674953 0.6719101 0.5027985 0.7931315 0.9491894
#>  [442] 0.9788155 0.9655996 0.9833517 0.9833517 0.5027985 0.9887244 0.7931315
#>  [449] 0.7262641 0.8185791 0.7990959 0.9266385 0.9799603 0.8131668 0.9576995
#>  [456] 0.5027985 0.5313247 0.9477132 0.8726056 0.9897774 0.9549011 0.9231711
#>  [463] 0.6676842 0.9576995 0.5027985 0.9844383 0.8158864 0.9729890 0.9908225
#>  [470] 0.9776599 0.9939125 0.9887244 0.9549011 0.9549011 0.9563075 0.6802830
#>  [477] 0.6145321 0.9576995 0.9908225 0.9549011 0.9729890 0.8264171 0.5027985
#>  [484] 0.6050033 0.9822393 0.9491894 0.8314871 0.8702972 0.9833517 0.9025417
#>  [491] 0.9383299 0.9729890 0.5027985 0.9822393 0.8726056 0.9681142 0.6376233
#>  [498] 0.5027985 0.9178243 0.8414829 0.9520570 0.9887244 0.5027985 0.9491894
#>  [505] 0.9776599 0.9630332 0.8264171 0.9729890 0.8131668 0.9833517 0.6761095
#>  [512] 0.9383299 0.7608503 0.7961472 0.9741627 0.9989950 0.9855177 0.6843811
#>  [519] 0.9178243 0.9908225 0.9283481 0.9383299 0.5899609 0.9617388 0.9928889
#>  [526] 0.9283481 0.8880489 0.9317363 0.8793472 0.8185791 0.5027985 0.9446998
#>  [533] 0.9741627 0.8609432 0.9617388 0.9249129 0.9446998 0.9833517 0.9520570
#>  [540] 0.9491894 0.9446998 0.6719101 0.9908225 0.9979828 0.9617388 0.9415296
#>  [547] 0.5027985 0.7990959 0.7961472 0.9693538 0.6843811 0.7708034 0.8264171
#>  [554] 0.6843811 0.9887244 0.8793472 0.8923287 0.9178243 0.9681142 0.9844383
#>  [561] 0.5638906 0.8075972 0.8985304 0.5027985 0.9908225 0.9788155 0.8964742
#>  [568] 0.9064278 0.5027985 0.9788155 0.9604049 0.7369971 0.9025417 0.8923287
#>  [575] 0.5027985 0.9025417 0.9764960 0.9681142 0.7541759 0.9822393 0.5027985
#>  [582] 0.9563075 0.8609432 0.8314871 0.9865912 0.5027985 0.8585517 0.9025417
#>  [589] 0.8985304 0.8923287 0.9025417 0.9989950 0.9897774 0.9477132 0.9822393
#>  [596] 0.9693538 0.8075972 0.7931315 1.0000000 0.8488573 0.8185791 0.8726056
#>  [603] 0.9833517 0.9729890 0.6050033 0.9939125 0.9576995 0.9491894 0.9908225
#>  [610] 0.9563075 0.5027985 0.9833517 0.9811011 0.9897774 0.9822393 0.9491894
#>  [617] 0.8585517 0.7868986 0.9415296 0.5027985 0.6550067 0.7001390 0.7773915
#>  [624] 0.9897774 0.9446998 0.5027985 0.9681142 0.6550067 0.8793472 0.9102891
#>  [631] 0.8748721 0.5027985 0.9563075 0.5027985 0.7740975 0.9928889 0.9350404
#>  [638] 0.9799603 0.7077701 0.9668585 0.9477132 0.9576995 0.8488573 0.7262641
#>  [645] 0.5533004 0.9788155 0.8815401 0.9776599 0.9989950 0.9300484 0.5027985
#>  [652] 0.9102891 0.9213957 0.5027985 0.9949319 0.5027985 0.9693538 0.9655996
#>  [659] 0.9491894 0.9833517 0.5027985 0.8837221 0.9102891 0.9383299 0.8656802
#>  [666] 0.8439506 0.9811011 0.9928889 0.8901939 0.9887244 0.8656802 0.9446998
#>  [673] 0.9693538 0.9446998 0.9178243 0.8390007 0.9822393 0.5586206 0.6050033
#>  [680] 0.7931315 0.9693538 0.9178243 0.9317363 0.8185791 0.7474120 0.9178243
#>  [687] 0.5479286 0.9415296 0.9383299 0.5027985 0.6923162 0.7152601 0.6962276
#>  [694] 0.7298418 0.8793472 0.5027985 0.5201364 0.7474120 0.9655996 0.9178243
#>  [701] 0.7405234 0.8238264 0.9491894 0.8185791 0.8656802 0.7990959 0.5027985
#>  [708] 0.9383299 0.9655996 0.9491894 0.9729890 0.9178243 0.9855177 0.5027985
#>  [715] 0.9446998 0.9317363 0.9383299 0.5479286 0.8985304 0.9140740 0.9477132
#>  [722] 0.9989950 0.9887244 0.9178243 0.5027985 0.9477132 0.8158864 0.5027985
#>  [729] 0.8880489 0.7507939 0.9491894 0.9717863 0.5201364 0.9717863 0.9897774
#>  [736] 0.9446998 0.7931315 0.9617388 0.9102891 0.9300484 0.9655996 0.9887244
#>  [743] 0.9655996 0.9317363 0.8414829 0.9833517 0.8923287 0.5027985 0.9196170
#>  [750] 0.9822393 0.7369971 0.8901939 0.9249129 0.5143571 0.5027985 0.8264171
#>  [757] 0.9491894 0.9477132 0.9576995 0.8656802 0.8923287 0.9102891 0.8944039
#>  [764] 0.8365037 0.9005405 0.9822393 0.9617388 0.9693538 0.9855177 0.8793472
#>  [771] 0.9604049 0.8726056 0.9833517 0.7115348 0.9249129 0.8289521 0.8726056
#>  [778] 0.9576995 0.9231711 0.9617388 0.9477132 0.7990959 0.7773915 0.5027985
#>  [785] 1.0000000 0.9822393 0.7805775 0.7990959 0.6050033 0.9477132 0.6050033
#>  [792] 0.9383299 0.9196170 0.7674953 0.8793472 0.8815401 0.9833517 0.9446998
#>  [799] 0.6843811 0.8488573 0.9231711 0.9446998 0.9462113 0.9887244 0.9855177
#>  [806] 0.9590579 0.9005405 0.9317363 0.9415296 0.7474120 0.8656802 0.9576995
#>  [813] 0.5027985 0.5586206 0.9741627 0.9693538 0.8212028 0.9822393 0.7541759
#>  [820] 0.9741627 0.9729890 0.9928889 0.9121816 0.9788155 0.9102891 0.9617388
#>  [827] 0.9196170 0.9928889 0.8656802 0.8439506 0.5027985 0.9668585 0.6098049
#>  [834] 0.8585517 0.8103820 0.9446998 0.9887244 0.9822393 0.9822393 0.9729890
#>  [841] 0.9939125 0.8488573 0.9617388 0.7152601 0.8901939 0.9446998 0.9025417
#>  [848] 0.8656802 0.5797007 0.9655996 0.7990959 0.9590579 0.5257901 0.8923287
#>  [855] 0.8726056 0.6192593 0.9025417 0.7115348 0.9729890 0.9231711 0.9693538
#>  [862] 0.6506974 0.9563075 0.5797007 0.9717863 0.9576995 0.5027985 0.8964742
#>  [869] 0.5027985 0.9399311 0.9630332 0.8858934 0.5027985 0.9025417 0.7405234
#>  [876] 0.9178243 0.9366866 0.9083605 0.9729890 0.5027985 0.9102891 0.9604049
#>  [883] 0.8656802 0.9844383 0.5027985 0.9446998 0.9897774 0.9887244 0.8985304
#>  [890] 0.9788155 0.5899609 0.8880489 0.7189281 0.9446998 0.9865912 0.9231711
#>  [897] 0.5027985 0.7900212 0.7773915 0.8131668 0.5027985 0.9590579 0.8858934
#>  [904] 0.7225961 0.8858934 0.9822393 0.8964742 0.7931315 0.9788155 0.8339992
#>  [911] 0.7405234 0.9979828 0.9908225 0.9939125 0.7152601 0.7837635 0.9833517
#>  [918] 0.9776599 0.9788155 0.9729890 0.9833517 0.9753307 0.9159626 0.9283481
#>  [925] 0.7961472 0.5027985 0.9333883 0.5950033 0.9865912 0.8964742 0.8923287
#>  [932] 0.9549011 0.9266385 0.9717863 0.8985304 0.9833517 0.5027985 0.9025417
#>  [939] 0.9231711 0.7334194 0.9729890 0.9865912 0.6883601 0.8585517 0.9383299
#>  [946] 0.9617388 0.9655996 0.7039649 0.8238264 0.5257901 0.9477132 0.9431147
#>  [953] 0.8131668 0.9908225 0.8439506 0.9764960 0.9822393 0.8561332 0.9025417
#>  [960] 0.6331953 0.5027985 0.9776599 0.9681142 0.5027985 0.9159626 0.7773915
#>  [967] 0.9590579 0.9446998 0.9693538 0.9231711 0.9196170 0.9643190 0.5027985
#>  [974] 0.7990959 0.6883601 0.5027985 0.9833517 0.9231711 0.9928889 0.9865912
#>  [981] 0.5085778 0.9753307 0.9383299 0.7575278 0.9729890 0.9350404 0.9576995
#>  [988] 0.5027985 0.9655996 0.9178243 0.9025417 0.9776599 0.9477132 0.9491894
#>  [995] 0.8985304 0.5027985 0.6192593 0.8923287 0.8985304 0.9939125
#> [1] "u:"
#>       Z[, 1]       Z[, 2]       Z[, 3] 
#> -0.140959342  0.001143142  0.137141524 
#> [1] "u:"
#>       Z[, 1]       Z[, 2]       Z[, 3] 
#> -0.008621646 -0.000022194  0.008483000 
#> [1] "u:"
#>        Z[, 1]        Z[, 2]        Z[, 3] 
#> -9.427849e-05  1.091706e-05  1.280852e-04 
#> [1] "u:"
#>       Z[, 1]       Z[, 2]       Z[, 3] 
#> 1.764027e-06 1.233501e-06 1.845976e-06 
#> [1] "u:"
#>        Z[, 1]        Z[, 2]        Z[, 3] 
#>  4.790446e-10 -1.582057e-10  4.893738e-10 
# Or you may generate another one
set.seed(1)
df = 10
nn = 1000
beta = c(0.5,0,-0.5, rep(0,10))
sim_PO_data = PO.sim(nn, beta,
                     C.gen = function(n)
                       5+rbinom(n,1,0.5)*runif(n, -5, 0))
# Fit PO model
res = PO(Surv(X, delta) ~ Z[,1]+ Z[,2]+ Z[,3]+ Z[,4]+ Z[,5]+ Z[,6],data = sim_PO_data)
#>    [1] 0.8748721 0.9928889 0.9102891 0.8793472 0.9693538 0.8019397 0.9969686
#>    [8] 0.9231711 0.9887244 0.6634584 0.7837635 0.9989950 0.5533004 0.9317363
#>   [15] 0.8609432 0.8131668 0.9604049 0.9178243 0.5027985 0.9563075 0.9855177
#>   [22] 0.6550067 0.9643190 0.6463882 0.6239506 0.9681142 0.7262641 0.9717863
#>   [29] 0.9617388 0.9822393 0.9617388 0.9383299 0.5257901 0.9159626 0.7115348
#>   [36] 0.8793472 0.9159626 0.9617388 0.9681142 0.8019397 0.9717863 0.9159626
#>   [43] 0.8075972 0.9989950 0.9681142 0.5027985 0.9446998 0.8537079 0.9446998
#>   [50] 0.9231711 0.8414829 0.8131668 0.9576995 0.9908225 0.8185791 0.9655996
#>   [57] 0.9729890 0.5027985 0.5027985 0.9350404 0.6463882 0.8075972 0.9446998
#>   [64] 0.9617388 0.9833517 0.7961472 0.9833517 0.7369971 0.8047734 0.7773915
#>   [71] 1.0000000 0.7990959 0.7439677 0.9576995 0.9520570 0.5027985 0.9064278
#>   [78] 0.9549011 0.5027985 0.9005405 0.9576995 0.9969686 0.9887244 0.9876578
#>   [85] 0.9534802 0.9717863 0.9908225 0.9491894 0.8923287 0.5027985 0.9753307
#>   [92] 0.9549011 0.9969686 0.9844383 0.7961472 0.9446998 0.9064278 0.9865912
#>   [99] 0.5027985 0.8923287 0.7837635 0.6843811 0.5027985 1.0000000 0.8464040
#>  [106] 0.9887244 0.8985304 0.9178243 0.9949319 0.9249129 0.8131668 0.9520570
#>  [113] 0.8609432 0.5691607 0.5201364 0.5027985 0.9604049 0.9969686 0.9833517
#>  [120] 0.7990959 0.9102891 0.8679887 0.5797007 0.6923162 0.9446998 0.9266385
#>  [127] 0.9549011 0.7837635 0.9383299 0.5027985 0.9178243 0.9729890 0.9693538
#>  [134] 0.6000033 0.9799603 0.6550067 0.6286070 0.7961472 0.7077701 0.8047734
#>  [141] 0.8837221 0.9617388 0.7708034 0.9928889 0.9655996 0.7405234 0.8985304
#>  [148] 0.9630332 0.8633149 0.9383299 0.9159626 0.6331953 0.9799603 0.7001390
#>  [155] 0.5027985 0.8314871 0.9908225 0.9576995 0.9266385 0.9462113 0.7405234
#>  [162] 0.7039649 0.5368593 0.8748721 0.9729890 0.5479286 0.7773915 0.8339992
#>  [169] 0.7541759 0.9520570 0.9462113 0.9668585 0.9399311 0.9681142 0.9655996
#>  [176] 0.9617388 0.9822393 0.9590579 0.9140740 0.9655996 0.5027985 0.9576995
#>  [183] 0.9693538 0.9520570 0.9643190 0.6761095 0.9897774 0.8264171 0.8633149
#>  [190] 0.7837635 0.8923287 0.9178243 0.9604049 0.9446998 0.9446998 0.5027985
#>  [197] 0.9617388 0.6802830 0.8264171 0.9822393 0.8837221 0.9604049 0.9693538
#>  [204] 1.0000000 0.7001390 0.9979828 0.9811011 0.9741627 0.9822393 0.9317363
#>  [211] 0.8158864 0.7262641 0.5027985 0.5848308 0.8858934 0.5027985 0.9617388
#>  [218] 0.6098049 0.9563075 0.8923287 0.8748721 0.9102891 0.9844383 0.5027985
#>  [225] 0.9383299 0.9446998 0.9928889 0.9025417 0.5027985 0.5027985 0.6286070
#>  [232] 0.8185791 0.9576995 0.9477132 0.7001390 0.9705700 0.9178243 0.9822393
#>  [239] 0.9729890 0.9491894 0.7262641 0.9822393 0.5797007 0.9491894 0.9729890
#>  [246] 0.7868986 0.7961472 0.9044868 0.9865912 0.9446998 0.9617388 0.9576995
#>  [253] 0.9266385 0.9969686 0.9844383 0.9717863 0.9918557 0.5027985 0.5027985
#>  [260] 0.9855177 0.9300484 0.9491894 0.7474120 0.8158864 0.9928889 0.9446998
#>  [267] 0.9044868 0.6050033 0.9617388 0.5027985 0.9908225 0.9317363 0.9897774
#>  [274] 0.7077701 0.9366866 0.7773915 0.5027985 0.8238264 0.6376233 0.9383299
#>  [281] 0.8944039 0.7001390 0.9590579 0.7773915 0.9178243 0.8264171 0.7900212
#>  [288] 0.9025417 0.7575278 0.9563075 0.9630332 0.9576995 0.9617388 0.9491894
#>  [295] 0.9317363 0.9159626 0.5950033 0.8390007 0.8488573 0.9617388 0.9300484
#>  [302] 0.6802830 1.0000000 0.6843811 0.8656802 0.9249129 0.6923162 0.9717863
#>  [309] 0.6098049 0.6331953 0.8656802 0.5027985 0.8726056 0.9643190 0.5027985
#>  [316] 0.9693538 0.9681142 0.5027985 0.9231711 0.7641728 0.9833517 0.9776599
#>  [323] 0.8985304 0.9317363 0.7369971 0.9178243 0.9693538 0.9729890 0.9249129
#>  [330] 0.5027985 0.9159626 0.9266385 0.6463882 0.9833517 0.9491894 0.9833517
#>  [337] 0.9844383 0.8185791 0.6802830 0.8365037 0.7931315 0.6286070 0.5027985
#>  [344] 0.9729890 0.8985304 0.6331953 0.5027985 0.8726056 0.9446998 0.9630332
#>  [351] 0.9317363 0.7262641 0.8264171 0.7474120 0.9617388 0.6420207 0.7773915
#>  [358] 0.9729890 0.8585517 0.9446998 0.8390007 0.7990959 0.5027985 0.9477132
#>  [365] 0.9681142 0.5744307 0.9534802 0.9446998 0.5027985 0.5027985 0.9178243
#>  [372] 0.9655996 0.7990959 0.9989950 0.5027985 0.9083605 0.9196170 0.9446998
#>  [379] 0.9415296 0.9462113 0.5423940 0.9630332 0.6331953 0.7931315 0.8512826
#>  [386] 0.9788155 0.8075972 0.9283481 0.8748721 0.5479286 0.8238264 0.7405234
#>  [393] 0.8815401 0.6331953 0.9178243 0.9317363 0.8185791 0.9415296 0.6592325
#>  [400] 0.9300484 0.9617388 0.9025417 0.7575278 0.6420207 0.7152601 0.9822393
#>  [407] 0.9491894 0.9213957 0.7931315 0.9590579 0.8748721 0.9462113 0.9491894
#>  [414] 0.9446998 0.6239506 0.8923287 0.5027985 0.9717863 0.5027985 0.9604049
#>  [421] 0.9317363 0.5027985 0.8238264 0.8314871 0.9959503 0.9833517 0.8561332
#>  [428] 0.9693538 0.8793472 0.9506232 0.9383299 0.5899609 0.5027985 0.8131668
#>  [435] 0.9989950 0.8771097 0.7674953 0.6719101 0.5027985 0.7931315 0.9491894
#>  [442] 0.9788155 0.9655996 0.9833517 0.9833517 0.5027985 0.9887244 0.7931315
#>  [449] 0.7262641 0.8185791 0.7990959 0.9266385 0.9799603 0.8131668 0.9576995
#>  [456] 0.5027985 0.5313247 0.9477132 0.8726056 0.9897774 0.9549011 0.9231711
#>  [463] 0.6676842 0.9576995 0.5027985 0.9844383 0.8158864 0.9729890 0.9908225
#>  [470] 0.9776599 0.9939125 0.9887244 0.9549011 0.9549011 0.9563075 0.6802830
#>  [477] 0.6145321 0.9576995 0.9908225 0.9549011 0.9729890 0.8264171 0.5027985
#>  [484] 0.6050033 0.9822393 0.9491894 0.8314871 0.8702972 0.9833517 0.9025417
#>  [491] 0.9383299 0.9729890 0.5027985 0.9822393 0.8726056 0.9681142 0.6376233
#>  [498] 0.5027985 0.9178243 0.8414829 0.9520570 0.9887244 0.5027985 0.9491894
#>  [505] 0.9776599 0.9630332 0.8264171 0.9729890 0.8131668 0.9833517 0.6761095
#>  [512] 0.9383299 0.7608503 0.7961472 0.9741627 0.9989950 0.9855177 0.6843811
#>  [519] 0.9178243 0.9908225 0.9283481 0.9383299 0.5899609 0.9617388 0.9928889
#>  [526] 0.9283481 0.8880489 0.9317363 0.8793472 0.8185791 0.5027985 0.9446998
#>  [533] 0.9741627 0.8609432 0.9617388 0.9249129 0.9446998 0.9833517 0.9520570
#>  [540] 0.9491894 0.9446998 0.6719101 0.9908225 0.9979828 0.9617388 0.9415296
#>  [547] 0.5027985 0.7990959 0.7961472 0.9693538 0.6843811 0.7708034 0.8264171
#>  [554] 0.6843811 0.9887244 0.8793472 0.8923287 0.9178243 0.9681142 0.9844383
#>  [561] 0.5638906 0.8075972 0.8985304 0.5027985 0.9908225 0.9788155 0.8964742
#>  [568] 0.9064278 0.5027985 0.9788155 0.9604049 0.7369971 0.9025417 0.8923287
#>  [575] 0.5027985 0.9025417 0.9764960 0.9681142 0.7541759 0.9822393 0.5027985
#>  [582] 0.9563075 0.8609432 0.8314871 0.9865912 0.5027985 0.8585517 0.9025417
#>  [589] 0.8985304 0.8923287 0.9025417 0.9989950 0.9897774 0.9477132 0.9822393
#>  [596] 0.9693538 0.8075972 0.7931315 1.0000000 0.8488573 0.8185791 0.8726056
#>  [603] 0.9833517 0.9729890 0.6050033 0.9939125 0.9576995 0.9491894 0.9908225
#>  [610] 0.9563075 0.5027985 0.9833517 0.9811011 0.9897774 0.9822393 0.9491894
#>  [617] 0.8585517 0.7868986 0.9415296 0.5027985 0.6550067 0.7001390 0.7773915
#>  [624] 0.9897774 0.9446998 0.5027985 0.9681142 0.6550067 0.8793472 0.9102891
#>  [631] 0.8748721 0.5027985 0.9563075 0.5027985 0.7740975 0.9928889 0.9350404
#>  [638] 0.9799603 0.7077701 0.9668585 0.9477132 0.9576995 0.8488573 0.7262641
#>  [645] 0.5533004 0.9788155 0.8815401 0.9776599 0.9989950 0.9300484 0.5027985
#>  [652] 0.9102891 0.9213957 0.5027985 0.9949319 0.5027985 0.9693538 0.9655996
#>  [659] 0.9491894 0.9833517 0.5027985 0.8837221 0.9102891 0.9383299 0.8656802
#>  [666] 0.8439506 0.9811011 0.9928889 0.8901939 0.9887244 0.8656802 0.9446998
#>  [673] 0.9693538 0.9446998 0.9178243 0.8390007 0.9822393 0.5586206 0.6050033
#>  [680] 0.7931315 0.9693538 0.9178243 0.9317363 0.8185791 0.7474120 0.9178243
#>  [687] 0.5479286 0.9415296 0.9383299 0.5027985 0.6923162 0.7152601 0.6962276
#>  [694] 0.7298418 0.8793472 0.5027985 0.5201364 0.7474120 0.9655996 0.9178243
#>  [701] 0.7405234 0.8238264 0.9491894 0.8185791 0.8656802 0.7990959 0.5027985
#>  [708] 0.9383299 0.9655996 0.9491894 0.9729890 0.9178243 0.9855177 0.5027985
#>  [715] 0.9446998 0.9317363 0.9383299 0.5479286 0.8985304 0.9140740 0.9477132
#>  [722] 0.9989950 0.9887244 0.9178243 0.5027985 0.9477132 0.8158864 0.5027985
#>  [729] 0.8880489 0.7507939 0.9491894 0.9717863 0.5201364 0.9717863 0.9897774
#>  [736] 0.9446998 0.7931315 0.9617388 0.9102891 0.9300484 0.9655996 0.9887244
#>  [743] 0.9655996 0.9317363 0.8414829 0.9833517 0.8923287 0.5027985 0.9196170
#>  [750] 0.9822393 0.7369971 0.8901939 0.9249129 0.5143571 0.5027985 0.8264171
#>  [757] 0.9491894 0.9477132 0.9576995 0.8656802 0.8923287 0.9102891 0.8944039
#>  [764] 0.8365037 0.9005405 0.9822393 0.9617388 0.9693538 0.9855177 0.8793472
#>  [771] 0.9604049 0.8726056 0.9833517 0.7115348 0.9249129 0.8289521 0.8726056
#>  [778] 0.9576995 0.9231711 0.9617388 0.9477132 0.7990959 0.7773915 0.5027985
#>  [785] 1.0000000 0.9822393 0.7805775 0.7990959 0.6050033 0.9477132 0.6050033
#>  [792] 0.9383299 0.9196170 0.7674953 0.8793472 0.8815401 0.9833517 0.9446998
#>  [799] 0.6843811 0.8488573 0.9231711 0.9446998 0.9462113 0.9887244 0.9855177
#>  [806] 0.9590579 0.9005405 0.9317363 0.9415296 0.7474120 0.8656802 0.9576995
#>  [813] 0.5027985 0.5586206 0.9741627 0.9693538 0.8212028 0.9822393 0.7541759
#>  [820] 0.9741627 0.9729890 0.9928889 0.9121816 0.9788155 0.9102891 0.9617388
#>  [827] 0.9196170 0.9928889 0.8656802 0.8439506 0.5027985 0.9668585 0.6098049
#>  [834] 0.8585517 0.8103820 0.9446998 0.9887244 0.9822393 0.9822393 0.9729890
#>  [841] 0.9939125 0.8488573 0.9617388 0.7152601 0.8901939 0.9446998 0.9025417
#>  [848] 0.8656802 0.5797007 0.9655996 0.7990959 0.9590579 0.5257901 0.8923287
#>  [855] 0.8726056 0.6192593 0.9025417 0.7115348 0.9729890 0.9231711 0.9693538
#>  [862] 0.6506974 0.9563075 0.5797007 0.9717863 0.9576995 0.5027985 0.8964742
#>  [869] 0.5027985 0.9399311 0.9630332 0.8858934 0.5027985 0.9025417 0.7405234
#>  [876] 0.9178243 0.9366866 0.9083605 0.9729890 0.5027985 0.9102891 0.9604049
#>  [883] 0.8656802 0.9844383 0.5027985 0.9446998 0.9897774 0.9887244 0.8985304
#>  [890] 0.9788155 0.5899609 0.8880489 0.7189281 0.9446998 0.9865912 0.9231711
#>  [897] 0.5027985 0.7900212 0.7773915 0.8131668 0.5027985 0.9590579 0.8858934
#>  [904] 0.7225961 0.8858934 0.9822393 0.8964742 0.7931315 0.9788155 0.8339992
#>  [911] 0.7405234 0.9979828 0.9908225 0.9939125 0.7152601 0.7837635 0.9833517
#>  [918] 0.9776599 0.9788155 0.9729890 0.9833517 0.9753307 0.9159626 0.9283481
#>  [925] 0.7961472 0.5027985 0.9333883 0.5950033 0.9865912 0.8964742 0.8923287
#>  [932] 0.9549011 0.9266385 0.9717863 0.8985304 0.9833517 0.5027985 0.9025417
#>  [939] 0.9231711 0.7334194 0.9729890 0.9865912 0.6883601 0.8585517 0.9383299
#>  [946] 0.9617388 0.9655996 0.7039649 0.8238264 0.5257901 0.9477132 0.9431147
#>  [953] 0.8131668 0.9908225 0.8439506 0.9764960 0.9822393 0.8561332 0.9025417
#>  [960] 0.6331953 0.5027985 0.9776599 0.9681142 0.5027985 0.9159626 0.7773915
#>  [967] 0.9590579 0.9446998 0.9693538 0.9231711 0.9196170 0.9643190 0.5027985
#>  [974] 0.7990959 0.6883601 0.5027985 0.9833517 0.9231711 0.9928889 0.9865912
#>  [981] 0.5085778 0.9753307 0.9383299 0.7575278 0.9729890 0.9350404 0.9576995
#>  [988] 0.5027985 0.9655996 0.9178243 0.9025417 0.9776599 0.9477132 0.9491894
#>  [995] 0.8985304 0.5027985 0.6192593 0.8923287 0.8985304 0.9939125
#> [1] "u:"
#>        Z[, 1]        Z[, 2]        Z[, 3]        Z[, 4]        Z[, 5] 
#> -0.1409593420  0.0011431418  0.1371415239 -0.0006879025  0.0063192958 
#>        Z[, 6] 
#> -0.0081719878 
#> [1] "u:"
#>        Z[, 1]        Z[, 2]        Z[, 3]        Z[, 4]        Z[, 5] 
#> -8.696130e-03  7.564393e-06  8.546967e-03 -1.785198e-04  4.094353e-04 
#>        Z[, 6] 
#> -4.472289e-04 
#> [1] "u:"
#>        Z[, 1]        Z[, 2]        Z[, 3]        Z[, 4]        Z[, 5] 
#> -1.143158e-04 -3.593012e-07  1.132146e-04 -3.515839e-06  5.290573e-06 
#>        Z[, 6] 
#> -5.449954e-06 
#> [1] "u:"
#>        Z[, 1]        Z[, 2]        Z[, 3]        Z[, 4]        Z[, 5] 
#> -2.073627e-08 -1.526551e-09  2.169916e-08 -8.091520e-10  1.395694e-09 
#>        Z[, 6] 
#> -1.358480e-10