NPMLE for Logistic-CoxPH Cure-Rate Model

NPMLE for Logistic-CoxPH Cure-Rate Model

PO(formula, data, C, df, weights, subset, init,control,
singular.ok = TRUE,model = FALSE,x = FALSE, y = TRUE, tt,
method = c('U-method','B-spline','NPMLE','glasso','glasso-PLH'),...)

Arguments

formula

a formula object, with the response on the left of a ~ operator, and the terms on the right. The response must be a survival object as returned by the Surv.cure function.

data

a data.frame in which to interpret the variables named in the formula, or in the subset and the weights argument.

C

need to edit.

df

degree of freedom

weights

vector of case weights, see the note below. For a thorough discussion of these see the book by Therneau and Grambsch.

subset

expression indicating which subset of the rows of data should be used in the fit. All observations are included by default.

init

vector of initial values of the iteration. Default initial value is zero for all variables.

control

Object of class PO.control specifying iteration limit and other control options. Default is PO.control(...).

singular.ok

logical value indicating how to handle collinearity in the model matrix. If TRUE, the program will automatically skip over columns of the X matrix that are linear combinations of earlier columns. In this case the coefficients for such columns will be NA, and the variance matrix will contain zeros. For ancillary calculations, such as the linear predictor, the missing coefficients are treated as zeros.

model

logical value: if TRUE, the model frame is returned in component model.

x

logical value: if TRUE, the x matrix is returned in component x.

y

logical value: if TRUE, the response vector is returned in component y.

tt

optional list of time-transform functions.

method

a character string specifying the method in c('U-method', 'B-spline','NPMLE') for estimation. The default method is the U-method.

...

other parameters passed to PO.control.

Author

Jue Hou

Examples

 # A simulated data set
require(survival)
data('sim_PO_data')
res = PO(Surv(X, delta) ~ Z[,1]+ Z[,2]+ Z[,3],data = sim_PO_data)
# Or you may generate another one
set.seed(1)
df = 10
nn = 1000
beta = c(0.5,0,-0.5, rep(0,10))
sim_PO_data = PO.sim(nn, beta,
                     C.gen = function(n)
                       5+rbinom(n,1,0.5)*runif(n, -5, 0))
# Fit PO model
res = PO(Surv(X, delta) ~ Z[,1]+ Z[,2]+ Z[,3]+ Z[,4]+ Z[,5]+ Z[,6],data = sim_PO_data)